Improve Macros
This changes the names and definitions of various macros around to be
more predictable. As a result, some sources needed to change to ensure
tests remain passing. I also fixed a possible memory leak in
assertions.cpp where the message wouldn't be freed if vformat_to threw.
It doesn't matter, but I care nonetheless.