MT-Safe Assertions
This adds a lot of functionality and fixes several issues. Most
importantly, it introduces thread safety features for assertions.
Basically, what this does is allow a program to fail assertions on
multiple threads. What happens is, when an assertion fails on one
thread, another failure happening on another thread can lock a mutex to
ensure its message is printed too. This ensures that roughly
simultaneous parallel failures are collected and reported before the
program aborts.

To support this, a static thread_local buffer is used for message
formatting instead of allocating strings. The benefit to this is that
allocation errors have to occur during the allocation of this buffer.
The downside is that overlong messages are truncated. The buffer is
controlled by build options but defaults to a size of 4 by 1000 bytes.
This should be enough for 1000 characters whatever language the client
application uses.

In addition, when the buffer size is set to 0 (e.g., the code point size
is 0 or the buffer size is 0 code points) custom formatted messages are
disabled. In this case, *_MSG* assertions act as-if they were basic
assertions.

Finally, special error handling code is added. This case handles
possible formatting or concurrency errors (i.e., errors locking a
mutex). When an error occurs the library does a best effort to write out
to stderr and explain what happened. The program will still be aborted.